Move kubectl logis to k8s package
