+++ cli/k8s/kubectl_test.go
@@ -1,4 +1,4 @@
-package shell
+package k8s
 
 import (
 	"testing"
